\name{LogConcDEAD-internal}\title{Internal log-concave maximum
  likelihood estimation functions}

\alias{initialf}
\alias{initialy}
\alias{mle.samples}
\alias{print.LogConcDEAD}
\alias{lcd.eval.point}
\alias{summary.LogConcDEAD}
\alias{convhullnew}
\alias{is.in}
\alias{integratemarg}
\alias{J00AD_appr}
\alias{J00AD_ord}
\alias{J00AD}
\alias{eval2}
\description{
  Internal functions for LogConcDEAD
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :)).
}

\keyword{ internal }
