\name{LogConcDEAD-internal}\title{Internal log-concave maximum
  likelihood estimation functions}

\alias{initialf}
\alias{initialy}
\alias{mle.samples}
\alias{lcd.eval.point}
\alias{convhullnew}
\alias{is.in}
\alias{integratemarg}
\alias{J00AD_appr}
\alias{J00AD_ord}
\alias{J00AD}
\alias{eval2}
\description{
  Internal functions for \pkg{LogConcDEAD}
}
\details{
  These are not to be called by the user.
}

\keyword{ internal }
