\name{LogConcDEAD-internal}\title{Internal log-concave maximum
  likelihood estimation functions}

\alias{initialf}
\alias{initialy}
\alias{convhullnew}
\alias{integratemarg}
\alias{JAD}
\alias{JAD.approx}
\alias{JAD.ord}
\description{
  Internal functions for \pkg{LogConcDEAD}
}
\details{
  These are not to be called by the user.
}

\keyword{ internal }
