\name{LogConcDEAD-internal}\title{Internal log-concave maximum
  likelihood estimation functions}

\alias{eval.est}
\alias{eval.est2}
\alias{eval.marg}
\alias{eval.marg2}
\alias{eval.vect}
\alias{eval.vect2}
\alias{find.limits}
\alias{find.limits2}
\alias{initialf}
\alias{initialy}
\alias{mle.samples}
\alias{print.LogConcDEAD}
\alias{print.LCDweights}
\alias{lcd.eval.point}
\alias{summary.LogConcDEAD}
\alias{convhullnew}

\description{
  Internal functions for LogConcDEAD
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :)).
}

\keyword{ internal }
